Struct transaction_emitter::TxnEmitter

pub struct TxnEmitter<'t, 'd> { /* fields omitted */ }

Implementations

impl<'t, 'd> TxnEmitter<'t, 'd>

pub fn new(
treasury_compliance_account: &'t mut LocalAccount,
designated_dealer_account: &'d mut LocalAccount,
client: RestClient,
transaction_factory: TransactionFactory,
rng: StdRng
) -> Self

pub fn take_account(&mut self) -> LocalAccount

pub fn clear(&mut self)

pub fn rng(&mut self) -> &mut StdRng

pub fn from_rng(&mut self) -> StdRng

pub async fn get_money_source(
&mut self,
coins_total: u64
) -> Result<&mut LocalAccount>

pub async fn load_vasp_account(
&self,
client: &RestClient,
index: usize
) -> Result<LocalAccount>

pub async fn get_seed_accounts(
&mut self,
rest_clients: &[RestClient],
seed_account_num: usize,
vasp: bool
) -> Result<Vec<LocalAccount>>

pub async fn mint_accounts(
&mut self,
req: &EmitJobRequest,
total_requested_accounts: usize
) -> Result<()>

workflow of mint accounts:

mint faucet account as the money source

load tc account to create seed accounts(parent VASP), one seed account for each endpoint

mint coins from faucet to new created seed accounts

split number of requested accounts(child VASP) into equally size of groups

each seed account take responsibility to create one size of group requested accounts and mint coins to them example: requested totally 100 new accounts with 10 endpoints will create 10 seed accounts(parent VASP), each seed account create 10 new accounts

pub async fn start_job(&mut self, req: EmitJobRequest) -> Result<EmitJob>

pub async fn stop_job(&mut self, job: EmitJob) -> TxnStats

pub fn peek_job_stats(&self, job: &EmitJob) -> TxnStats

pub async fn periodic_stat(
&mut self,
job: &EmitJob,
duration: Duration,
interval_secs: u64
)

pub async fn emit_txn_for(
&mut self,
duration: Duration,
emit_job_request: EmitJobRequest
) -> Result<TxnStats>

pub async fn emit_txn_for_with_stats(
&mut self,
duration: Duration,
emit_job_request: EmitJobRequest,
interval_secs: u64
) -> Result<TxnStats>

pub async fn submit_single_transaction(
&self,
client: &RestClient,
sender: &mut LocalAccount,
receiver: &AccountAddress,
num_coins: u64
) -> Result<Instant>